UT: export credential testing infrastructure to conftest
